// src/app/modules/core/direct-business/models/doiModel.ts
export class AttachmentTypeList {
    attachmentType: string;
    type?: string;
}

export class ClaimEntryListingApplicationAccepted {
    policyNo: string;
    insuredName: string;
    policyType: string;
    district: string;
    taluka: string;
    claimEnteredDate: Date;
    claimAmount: Number;
    status: string;
}
// ...
export class TalukaList {
    value: string;
    district: string;
    viewValue: string;
}